move model out of lib
