+++ packages/frontend/web/server/document.tsx
@@ -5,7 +5,7 @@ import { renderToString } from 'react-dom/server';
 import htmlTemplate from './htmlTemplate';
 import Article from '../pages/Article';
 import assets from '@frontend/lib/assets';
-import { GADataType } from '@frontend/lib/model/extract-ga';
+import { GADataType } from '@frontend/model/extract-ga';
 
 interface Props {
     data: {
